FR-Arcface-Onnx/requirements.txt
onnx
mxnet-cu90mkl
numpy
matplotlib
opencv-python
scikit-learn
scikit-image
easydict
